babl-fish-reference: do some conversions in single precision float
This permits these conversions to go through floating point instead of double
precision when non of the involved format have more than 32bit precision,
this might be determinetal to u32 but is right for u8 and u16 which we rely
most on.

We also disable the missing fast path messages, since now many references;
through internal shortcuts are faster than some long chains, perhaps some
of these reference fast paths should be split out as new categories of fishes.

+++ babl/base/type-u16.c
@@ -112,8 +112,8 @@ MAKE_CONVERSIONS (u16, 0.0, 1.0, 0, UINT16_MAX)
 
 static inline void
 convert_float_u16_scaled (BablConversion *conversion,
-                          float    min_val,
-                          float    max_val,
+                          double    min_val,
+                          double    max_val,
                           uint16_t min,
                           uint16_t max,
                           char    *src,
@@ -142,8 +142,8 @@ convert_float_u16_scaled (BablConversion *conversion,
 
 static inline void
 convert_u16_float_scaled (BablConversion *conversion,
-                          float   min_val,
-                          float   max_val,
+                          double   min_val,
+                          double   max_val,
                           uint16_t min,
                           uint16_t max,
                           char    *src,
